+++ src/controllers/EventHandler.js
@@ -1,0 +1,8 @@
+const EventHandler = {
+  modalClose: (formElement, mainElement) => {
+    formElement.reset();
+    mainElement.querySelector(".modal").classList.remove("modal--open");
+  },
+};
+
+export default EventHandler;
